backend/scrapers/industry.py
import requests
from bs4 import BeautifulSoup
from typing import List

def search_industry_trends(domain: str) -> List[str]:
    """
    Search for industry trends in a given domain.
    For demonstration, we wrap a search query.
    In a production app, this might use SerpApi or similar.
    """
    # Placeholder for real searching logic
    # Real logic would scrape sites like LinkedIn, Glassdoor, or tech blogs
    # For now, let's return some realistic-looking trends based on the domain
    
    trends = {
        "Artificial Intelligence": ["Generative AI", "LLM Ops", "Vector Databases", "Prompt Engineering"],
        "Data Science": ["Data Engineering", "Machine Learning in Production", "Ethical AI", "AutoML"],
        "Software Engineering": ["Microservices", "Cloud Native", "DevSecOps", "Platform Engineering"]
    }
    
    return trends.get(domain, ["Modern techniques", "Emerging technologies", "Industry standards"])

def get_job_market_skills(domain: str) -> List[str]:
    """
    Scrape job market skills for a domain.
    """
    # This would ideally use an API or scrape a job board.
    return ["Collaboration", "Agile Methodology", "Problem Solving"]
