backend/requirements.txt
fastapi
pydantic
langgraph
langchain-core
langchain-openai
llama-index
llama-index-vector-stores-chroma
chromadb
beautifulsoup4
playwright
python-dotenv
llama-index-llms-openai
requests
pypdf
reportlab
python-multipart
gunicorn
uvicorn[standard]
